+++ .eslintrc.js
@@ -1,7 +1,8 @@
 module.exports = {
     'env': {
         'browser': true,
-        'es6': true
+        'es6': true,
+        'node': true
     },
     'extends': [
         'eslint:recommended',
@@ -21,6 +22,6 @@ module.exports = {
         'vue'
     ],
     'rules': {
-        'quotes': ['error', 'single'],
+        'quotes': ['error', 'single']
     }
 };