+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### Release 0.7.20-beta 
 
 ### feature 
-- updated mew-select to get the currency symbol and price. updated mew-table to truncate text for token column. fixed loading knob on mew-table storybook preview. 
+- removed static price symbol from mew-select. updated mew-table to truncate text for token column. fixed loading knob on mew-table storybook preview. 
 
 ### Release 0.7.19-beta